pyproject.toml
[project]
name = "llm-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "datasets>=4.6.1",
    "deepeval>=3.8.9",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
    "transformers>=5.3.0",
]
